requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
pyyaml>=6.0
streamlit>=1.28
jsonschema>=4.0
